Card 컴포넌트분리 (오류 하나 있는거 같아서 브랜치      해서 확인하기 전에 백업..)
