+++ src/pages/employer/WorkerManagePage.jsx
@@ -3,12 +3,12 @@ import { useNavigate } from "react-router-dom";
 import { FaUser, FaTimes } from "react-icons/fa";
 import Swal from "sweetalert2";
 import "../../styles/workerManagePage.css";
-import { formatKRW } from "../../utils/formatUtils";
-import EmployerTimeInput from "./components/EmployerTimeInput";
 import EmployerBasicInfoCard from "./components/EmployerBasicInfoCard";
 import EmployerWorkplaceForm from "./components/EmployerWorkplaceForm";
 import EmployerWorkInfoCard from "./components/EmployerWorkInfoCard";
 import EmployerScheduleGrid from "./components/EmployerScheduleGrid";
+import EmployerWorkerSearchCard from "./components/EmployerWorkerSearchCard";
+import EmployerNewWorkerWorkInfoCard from "./components/EmployerNewWorkerWorkInfoCard";
 import {
   getWorkplaces,
   createWorkplace,
@@ -1188,390 +1188,24 @@ export default function WorkerManagePage() {
           />
         ) : isAddingWorker ? (
           <>
-            {/* 근무자 코드 검색 카드 */}
-            <div className="info-card">
-              <div className="info-card-header">
-                <h3 className="info-card-title">근무자 코드</h3>
-              </div>
-              <div className="info-card-content">
-                <div className="info-field">
-                  <label className="info-label">근무자 코드</label>
-                  <div className="search-input-group">
-                    <input
-                      type="text"
-                      className="worker-code-input"
-                      value={workerCode}
-                      onChange={(e) => setWorkerCode(e.target.value)}
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter") {
-                          handleSearchWorker();
-                        }
-                      }}
-                      placeholder="근무자 코드를 입력하세요"
-                    />
-                    <button
-                      type="button"
-                      className="search-button"
-                      onClick={handleSearchWorker}
-                      disabled={isSearching}
-                    >
-                      {isSearching ? "검색 중..." : "검색"}
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
+            <EmployerWorkerSearchCard
+              workerCode={workerCode}
+              onWorkerCodeChange={setWorkerCode}
+              onSearch={handleSearchWorker}
+              searchedWorker={searchedWorker}
+              onConfirm={handleConfirmWorker}
+              isSearching={isSearching}
+            />
 
-            {/* 근무자 정보 확인 카드 */}
-            {searchedWorker && (
-              <div className="info-card">
-                <div className="info-card-header">
-                  <h3 className="info-card-title">근무자 정보</h3>
-                </div>
-                <div className="info-card-content">
-                  <div className="basic-info-header">
-                    <div className="profile-icon">
-                      <FaUser />
-                    </div>
-                    <div className="worker-info-display">
-                      <div className="info-field">
-                        <label className="info-label">이름</label>
-                        <div className="info-value">{searchedWorker.name}</div>
-                      </div>
-                      <div className="info-field">
-                        <label className="info-label">전화번호</label>
-                        <div className="info-value">
-                          {searchedWorker.phone || "-"}
-                        </div>
-                      </div>
-                      <div className="info-field">
-                        <label className="info-label">근무자 코드</label>
-                        <div className="info-value">
-                          {searchedWorker.workerCode}
-                        </div>
-                      </div>
-                    </div>
-                    <div className="confirm-buttons">
-                      <button
-                        type="button"
-                        className="confirm-button"
-                        onClick={handleConfirmWorker}
-                      >
-                        확인
-                      </button>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {/* 근무 정보 설정 카드 */}
             {confirmedWorker && newWorkerWorkInfo && (
-              <div className="info-card">
-                <div className="info-card-header">
-                  <h3 className="info-card-title">근무 정보</h3>
-                </div>
-                <div className="info-card-content">
-                  <div className="info-field">
-                    <label className="info-label">근무지</label>
-                    <div className="info-value">{selectedWorkplace}</div>
-                  </div>
-
-                  {/* 근무 시간 설정 */}
-                  <div className="info-field">
-                    <label className="info-label">근무 시간</label>
-                    <div className="weekly-schedule-inputs">
-                      {daysOfWeek.map((day) => {
-                        const schedule =
-                          newWorkerWorkInfo.weeklySchedule?.[day];
-                        return (
-                          <div key={day} className="day-schedule-row">
-                            <span className="day-label-small">{day}요일</span>
-                            <div className="time-wheel-wrapper">
-                              {schedule ? (
-                                <>
-                                  <EmployerTimeInput
-                                    value={schedule.start || "00:00"}
-                                    onChange={(val) => {
-                                      setNewWorkerWorkInfo({
-                                        ...newWorkerWorkInfo,
-                                        weeklySchedule: {
-                                          ...newWorkerWorkInfo.weeklySchedule,
-                                          [day]: {
-                                            ...schedule,
-                                            start: val,
-                                          },
-                                        },
-                                      });
-                                    }}
-                                  />
-                                  <span className="time-separator">~</span>
-                                  <EmployerTimeInput
-                                    value={schedule.end || "00:00"}
-                                    onChange={(val) => {
-                                      setNewWorkerWorkInfo({
-                                        ...newWorkerWorkInfo,
-                                        weeklySchedule: {
-                                          ...newWorkerWorkInfo.weeklySchedule,
-                                          [day]: {
-                                            ...schedule,
-                                            end: val,
-                                          },
-                                        },
-                                      });
-                                    }}
-                                    allowMidnight
-                                  />
-                                  {(() => {
-                                    const [startHour, startMin] = (
-                                      schedule.start || "00:00"
-                                    )
-                                      .split(":")
-                                      .map(Number);
-                                    const [endHour, endMin] = (
-                                      schedule.end || "00:00"
-                                    )
-                                      .split(":")
-                                      .map(Number);
-                                    const startDecimal =
-                                      startHour + startMin / 60;
-                                    const endDecimal = endHour + endMin / 60;
-                                    const crossesMidnight =
-                                      endDecimal <= startDecimal;
-                                    return crossesMidnight ? (
-                                      <span className="overnight-label">
-                                        (익일)
-                                      </span>
-                                    ) : null;
-                                  })()}
-                                  <button
-                                    type="button"
-                                    className="remove-schedule-button-x"
-                                    onClick={() => {
-                                      const updatedSchedule = {
-                                        ...newWorkerWorkInfo.weeklySchedule,
-                                      };
-                                      delete updatedSchedule[day];
-                                      setNewWorkerWorkInfo({
-                                        ...newWorkerWorkInfo,
-                                        weeklySchedule: updatedSchedule,
-                                      });
-                                    }}
-                                    title="삭제"
-                                  >
-                                    <FaTimes />
-                                  </button>
-                                </>
-                              ) : (
-                                <button
-                                  type="button"
-                                  className="add-schedule-button"
-                                  onClick={() => {
-                                    setNewWorkerWorkInfo({
-                                      ...newWorkerWorkInfo,
-                                      weeklySchedule: {
-                                        ...newWorkerWorkInfo.weeklySchedule,
-                                        [day]: { start: "09:00", end: "18:00" },
-                                      },
-                                    });
-                                  }}
-                                >
-                                  근무 추가
-                                </button>
-                              )}
-                            </div>
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-
-                  {/* 휴게 시간 설정 */}
-                  <div className="info-field">
-                    <label className="info-label">휴게 시간</label>
-                    <div className="break-time-by-day">
-                      {daysOfWeek.map((day) => {
-                        const breakTime =
-                          typeof newWorkerWorkInfo.breakTime === "object"
-                            ? newWorkerWorkInfo.breakTime[day] || 0
-                            : newWorkerWorkInfo.breakTime || 0;
-                        const hasSchedule =
-                          newWorkerWorkInfo.weeklySchedule?.[day];
-
-                        return (
-                          <div key={day} className="break-time-day-row">
-                            <span className="day-label-small">{day}요일</span>
-                            {hasSchedule ? (
-                              <div className="break-time-input-group">
-                                <input
-                                  type="number"
-                                  className="break-time-input-field"
-                                  value={breakTime === 0 ? "" : breakTime}
-                                  min="0"
-                                  onFocus={(e) => {
-                                    if (breakTime !== 0) {
-                                      e.target.select();
-                                    }
-                                  }}
-                                  onBlur={(e) => {
-                                    if (e.target.value === "") {
-                                      const newBreakTime =
-                                        typeof newWorkerWorkInfo.breakTime ===
-                                        "object"
-                                          ? { ...newWorkerWorkInfo.breakTime }
-                                          : daysOfWeek.reduce((acc, d) => {
-                                              acc[d] =
-                                                newWorkerWorkInfo.breakTime ||
-                                                0;
-                                              return acc;
-                                            }, {});
-                                      newBreakTime[day] = 0;
-                                      setNewWorkerWorkInfo({
-                                        ...newWorkerWorkInfo,
-                                        breakTime: newBreakTime,
-                                      });
-                                    }
-                                  }}
-                                  onClick={(e) => {
-                                    if (breakTime !== 0) {
-                                      e.target.select();
-                                    }
-                                  }}
-                                  onChange={(e) => {
-                                    const newBreakTime =
-                                      typeof newWorkerWorkInfo.breakTime ===
-                                      "object"
-                                        ? { ...newWorkerWorkInfo.breakTime }
-                                        : daysOfWeek.reduce((acc, d) => {
-                                            acc[d] =
-                                              newWorkerWorkInfo.breakTime || 0;
-                                            return acc;
-                                          }, {});
-                                    const inputValue = e.target.value;
-                                    newBreakTime[day] =
-                                      inputValue === ""
-                                        ? 0
-                                        : parseInt(inputValue, 10) || 0;
-                                    setNewWorkerWorkInfo({
-                                      ...newWorkerWorkInfo,
-                                      breakTime: newBreakTime,
-                                    });
-                                  }}
-                                />
-                                <span className="break-time-unit">분</span>
-                              </div>
-                            ) : (
-                              <span className="break-time-off">-</span>
-                            )}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-
-                  {/* 시급 및 급여 지급일 */}
-                  <div className="info-field-row">
-                    <div className="info-field">
-                      <label className="info-label">시급</label>
-                      <div className="wage-input-group">
-                        <input
-                          type="number"
-                          className="info-input"
-                          value={newWorkerWorkInfo.hourlyWage ?? ""}
-                          min="0"
-                          onChange={(e) =>
-                            setNewWorkerWorkInfo({
-                              ...newWorkerWorkInfo,
-                              hourlyWage:
-                                e.target.value === ""
-                                  ? null
-                                  : parseInt(e.target.value) || 0,
-                            })
-                          }
-                        />
-                        <span className="wage-unit">원</span>
-                      </div>
-                    </div>
-
-                    <div className="info-field">
-                      <label className="info-label">급여 지급일</label>
-                      <div className="payday-input-group">
-                        <span className="payday-text">매월</span>
-                        <input
-                          type="number"
-                          className="payday-input"
-                          value={newWorkerWorkInfo.payday ?? ""}
-                          min="1"
-                          max="31"
-                          onChange={(e) =>
-                            setNewWorkerWorkInfo({
-                              ...newWorkerWorkInfo,
-                              payday:
-                                e.target.value === ""
-                                  ? null
-                                  : parseInt(e.target.value) || 1,
-                            })
-                          }
-                        />
-                        <span>일</span>
-                      </div>
-                    </div>
-                  </div>
-
-                  {/* 4대보험 및 소득세 */}
-                  {/* <div className="toggle-row">
-                    <div className="toggle-item">
-                      <label className="toggle-label">4대 보험</label>
-                      <label className="toggle-switch">
-                        <input
-                          type="checkbox"
-                          checked={newWorkerWorkInfo.socialInsurance || false}
-                          onChange={(e) =>
-                            setNewWorkerWorkInfo({
-                              ...newWorkerWorkInfo,
-                              socialInsurance: e.target.checked,
-                            })
-                          }
-                        />
-                        <span className="toggle-slider"></span>
-                      </label>
-                    </div>
-                    <div className="toggle-item">
-                      <label className="toggle-label">소득세</label>
-                      <label className="toggle-switch">
-                        <input
-                          type="checkbox"
-                          checked={newWorkerWorkInfo.withholdingTax || false}
-                          onChange={(e) =>
-                            setNewWorkerWorkInfo({
-                              ...newWorkerWorkInfo,
-                              withholdingTax: e.target.checked,
-                            })
-                          }
-                        />
-                        <span className="toggle-slider"></span>
-                      </label>
-                    </div>
-                  </div> */}
-
-                  <div className="add-worker-button-container">
-                    <button
-                      type="button"
-                      className="cancel-button"
-                      onClick={handleCancelAddWorker}
-                    >
-                      취소
-                    </button>
-                    <button
-                      type="button"
-                      className="add-button-large"
-                      onClick={handleSaveNewWorker}
-                    >
-                      추가
-                    </button>
-                  </div>
-                </div>
-              </div>
+              <EmployerNewWorkerWorkInfoCard
+                confirmedWorker={confirmedWorker}
+                workInfo={newWorkerWorkInfo}
+                onWorkInfoChange={setNewWorkerWorkInfo}
+                onCancel={handleCancelAddWorker}
+                onSave={handleSaveNewWorker}
+                selectedWorkplace={selectedWorkplace}
+              />
             )}
           </>
         ) : workerData ? (
